backend/Code/requirements.txt
ffmpeg-python
openai-whisper
pysrt
pydub